feat(server): move goals schema out of routes
